pyproject.toml
[tool.poetry]
name = "web-py-ci"
version = "0.1.0"
description = "This project explores setting up CI pipelines for web application using Python, Docker, GitHub Actions, Travis CI, and CircleCI, offering a foundation for understanding CI automation with different tools."
authors = ["Deepika <dejain4@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
redis = "^5.0.2"



[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
isort = "^5.13.2"
pre-commit = "^3.6.2"
bandit = "^1.7.7"
coverage = "^7.4.3"



[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
